More test cases. Clean up autoconfig: use 64-bit offsets if possible, use GCC warnings if available. Be more fussy about using the right integer types. Clean up test directories on successful completion. Add -D and -x options to test driver script; make sure it is executable on install. Add ':' recognition to hsmapread. Add more map test strategies to genmaptest. Refactor mapptr code. Add test-pairs: pairs of files filtered through various edits.

+++ emit.c
@@ -225,7 +225,7 @@ _hs_emit_copy(hs_write_fn_t write_fn, void *write_priv,
     } else if (off_type == 4) {
 	cmd = op_copy_int_byte;
     } else {
-	_hs_fatal("can't pack offset %ld!", offset);
+	_hs_fatal("can't pack offset %ld!", (long) offset);
     }
 
     if (len_type == 1) {
